File changes and vježbe 13 contract

+++ vjezba12/ogranicen_pristup.sol
@@ -1,16 +1,20 @@
 pragma solidity ^0.8.0;
+
 // SPDX-License-Identifier: MIT
 
 contract ogranicen_pristup {
     bytes32 pass_hash;
-    int stanje;
+    int256 stanje;
 
-    constructor(string memory pass){
+    constructor(string memory pass) {
         pass_hash = sha256(bytes(pass));
         stanje = 500;
     }
 
-    function promijeni_pass(string memory pass, string memory novi_pass) public returns (bool){
+    function promijeni_pass(string memory pass, string memory novi_pass)
+        public
+        returns (bool)
+    {
         if (pass_hash == sha256(bytes(pass))) {
             pass_hash = sha256(bytes(novi_pass));
             return true;
@@ -19,8 +23,8 @@ contract ogranicen_pristup {
         }
     }
 
-    function povecajStanje(string memory pass) public returns (bool){
-        if (pass_hash == sha256(bytes(pass))){
+    function povecajStanje(string memory pass) public returns (bool) {
+        if (pass_hash == sha256(bytes(pass))) {
             stanje = stanje + 1;
             return true;
         } else {
@@ -28,8 +32,8 @@ contract ogranicen_pristup {
         }
     }
 
-    function smanjiStanje(string memory pass) public returns (bool){
-        if (pass_hash == sha256(bytes(pass))){
+    function smanjiStanje(string memory pass) public returns (bool) {
+        if (pass_hash == sha256(bytes(pass))) {
             stanje = stanje - 1;
             return true;
         } else {
@@ -37,8 +41,12 @@ contract ogranicen_pristup {
         }
     }
 
-    function dohvatiStanje(string memory pass) public view returns (int, bool){
-        if (pass_hash == sha256(bytes(pass))){
+    function dohvatiStanje(string memory pass)
+        public
+        view
+        returns (int256, bool)
+    {
+        if (pass_hash == sha256(bytes(pass))) {
             return (stanje, true);
         } else {
             return (0, false);
